backend/requirements.txt
fastapi
fastapi-cli
uvicorn
line-bot-sdk
requests
psycopg2-binary
pydantic
pydantic_core
python-dotenv
bcrypt
